make sound working again
Lol, it was most stupid bug that I fixed. I think the problem was with
close() function that was declared in init.c/.h files. It was clashing
with default close() function on unix systems for closing file
descriptors. But game compiled silently (no warnings about close) and crashed
at runtime while trying to init audio (some error with file descriptors).
P.S. I wrote this part of code with close() function when was using Windows,
so it was working at that time.